chore: move navigation menu components into dedicated directory
