+++ src/components/DashboardsBarNew/NavigationMenu/index.js
@@ -1,0 +1,2 @@
+export { IconNavigation } from './IconNavigation.js'
+export { NavigationMenu } from './NavigationMenu.js'
